Huge refactor homepage into modular components for improved readability and maintainability
