+++ src/app/page.tsx
@@ -1,406 +1,27 @@
 "use client";
 
 import React from "react";
-import Image from "next/image";
-import {
-  CalendarDays,
-  MapPin,
-  Users,
-  Megaphone,
-  Building,
-  ChevronRight,
-  Instagram,
-  Linkedin
-} from "lucide-react";
-import { Button } from "@/components/ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle
-} from "@/components/ui/card";
-import Sponsors from "@/components/sponsors";
-import Link from "next/link";
 import Nav from "@/components/nav";
-import FAQaccordion from "@/components/faq-accordion";
-import {
-  Carousel,
-  CarouselContent,
-  CarouselItem,
-  CarouselNext,
-  CarouselPrevious
-} from "@/components/ui/carousel";
-import StatBanner from "@/components/stat-banner";
-import { providers } from "../../public/cloud-providers/clouds";
-import Script from "next/script";
+import { HeroSection } from "@/components/hero-section";
+import { HighlightsSection } from "@/components/highlights-section";
+import { SponsorsSection } from "@/components/sponsors-section";
+import { CommunitySection } from "@/components/community-section";
+import { VenueSection } from "@/components/venue-section";
+import { FAQSection } from "@/components/faq-section";
 import Footer from "@/components/footer";
 
 export default function Home() {
-  // const openLink = () => {
-  //   window.open("https://www.vtixonline.com/cloud-ai-summit/4527/", "_blank");
-  // };
-
-  const openSpeakersLink = () => {
-    window.open("https://forms.gle/6qjgftM5Uf4ZSNNP7", "_blank");
-  };
-
-  const openVolunteersLink = () => {
-    window.open("https://forms.gle/NECDLqn6T6qbmWXZ8", "_blank");
-  };
-
   return (
     <div className="min-h-screen bg-black text-white">
-      {/* Main Banner */}
       <Nav />
-      <div className="relative min-h-[900px] flex flex-col items-center justify-center overflow-hidden  w-full bg-[#070B14] px-4 py-16 text-white">
-        <div className="absolute inset-0 z-0">
-          <div className="absolute inset-0 bg-gradient-to-r from-black/60 to-black/60"></div>
-          <video
-            src="/main-banner/highlight-video.mp4"
-            autoPlay
-            muted
-            loop
-            className="mix-blend-overlay object-cover h-full w-full"
-          />
-        </div>
-
-        <div className="relative mx-auto flex flex-row flex-wrap justify-center gap-6">
-          {/* Logo and Subtitle */}
-          <div className="md:mb-8 text-center">
-            <h1 className="mb-4 hidden md:block font-bold tracking-tight md:text-6xl lg:text-7xl">
-              <span className="bg-gradient-to-r from-green-900 to-green-500 px-6">
-                Cloud Summit
-              </span>
-            </h1>
-            <h1 className="mb-4 text-7xl font-bold tracking-tight md:hidden flex flex-wrap justify-center gap-2">
-              <span className="bg-gradient-to-r from-green-900 to-green-500 px-6 ">
-                Cloud
-              </span>
-              <span className="bg-gradient-to-r from-green-900 to-green-500 px-6 ">
-                Summit
-              </span>
-            </h1>
-            {/* <p className="text-xl md:text-2xl mb-8 font-bold text-gray-100">
-              ALL THE CLOUDS, ALL AT ONCE.
-            </p> */}
-            <h2 className="text-xl font-light md:text-2xl hidden md:block">
-              <span>Infrastructure</span>
-              <span className="mx-3 text-gray-500">|</span>
-              <span>Security</span>
-              <span className="mx-3 text-gray-500">|</span>
-              <span>DevOps</span>
-            </h2>
-          </div>
-
-          {/* Event Dates */}
-          <div className="mb-8 flex flex-col items-center md:items-start justify-start -mt-1">
-            <div className="flex items-center gap-4 mb-3 rounded-full bg-gradient-to-r from-green-900 to-green-500 px-6 py-2">
-              <span className="font-bold">May 27, 2025</span>
-              <span className="text-sm">Orpheum Theatre</span>
-            </div>
-            <div className="flex items-center gap-4 rounded-full border  border-white/20 px-6 py-2">
-              <span className="font-bold">Vancouver, BC</span>
-              <span className="text-sm">Canada</span>
-            </div>
-          </div>
-        </div>
-        <div className="relative mx-auto flex flex-col flex-wrap justify-center gap-6">
-          {/* Description */}
-          <p className="mx-auto mb-8 max-w-3xl text-center text-lg text-gray-300 md:text-xl">
-            Join cloud developers, architects, IT & infrastructure professionals
-            and executives building the cloud ecosystem.
-          </p>
-
-          {/* CTA Buttons */}
-          <div className="mb-2 md:mb-16 flex flex-col items-center justify-center gap-4 md:flex-row">
-            {/* <Button
-              size="lg"
-              className="min-w-[200px] bg-gradient-to-r from-green-900 to-green-500 font-bold"
-              onClick={openLink}
-            >
-              <span className="text">GET TICKETS</span>
-              <svg
-                className="next-arrow"
-                aria-hidden="false"
-                width="20px"
-                height="25px"
-                viewBox="0 0 50 80"
-                xmlSpace="preserve"
-              >
-                <polyline
-                  stroke="#ffffff"
-                  stroke-width="9"
-                  fill="none"
-                  stroke-linecap="round"
-                  stroke-linejoin="round"
-                  points="0, 0 45, 40 0, 80"
-                ></polyline>
-              </svg>
-            </Button> */}
-            <a
-              href="https://lu.ma/event/evt-cItbLfgBkf8na4n"
-              className="min-w-[200px] h-11 flex justify-center items-center rounded-md bg-gradient-to-r from-green-900 to-green-500 font-bold"
-              data-luma-action="checkout"
-              data-luma-event-id="evt-cItbLfgBkf8na4n"
-            >
-              Get Tickets
-            </a>
-
-            <Script
-              id="luma-checkout"
-              src="https://embed.lu.ma/checkout-button.js"
-              strategy="lazyOnload"
-            />
-
-            <Button
-              size="lg"
-              variant="outline"
-              className="min-w-[200px] border-white/20 text-white hover:bg-white/10 hidden md:block"
-              onClick={openVolunteersLink}
-            >
-              GET INVOLVED
-            </Button>
-          </div>
-          {/* Stats */}
-          <StatBanner />
-        </div>
-        <div className="grid grid-cols-4 gap-4 mt-4 md:mt-16 md:grid-cols-4 md:flex md:flex-row md:overflow-hidden md:whitespace-nowrap">
-          {providers.map((provider) => (
-            <a
-              href={provider.link}
-              key={provider.name}
-              target="_blank"
-              className="relative"
-            >
-              <img
-                src={provider.icon}
-                alt={provider.name}
-                className="w-14 h-auto object-cover"
-              />
-            </a>
-          ))}
-        </div>
-      </div>
-      {/* <header className="relative h-[75vh] flex flex-col items-center justify-center overflow-hidden">
-        <div className="absolute inset-0 z-0">
-          <div className="absolute inset-0 bg-gradient-to-r from-black/60 to-black/60"></div>
-          <video
-            src="/main-banner/highlight-video.mp4"
-            autoPlay
-            muted
-            loop
-            className="mix-blend-overlay object-cover h-[75vh] w-full"
-          />
-        </div>
-        <div className="relative z-10 text-center max-w-5xl mx-auto px-4">
-          <h1 className="text-4xl md:text-8xl font-bold mb-4 bg-clip-text text-transparent bg-gradient-to-r from-green-400 to-blue-500">
-            Cloud Summit 2025
-          </h1>
-          <p className="text-xl md:text-4xl mb-8 font-bold text-gray-100">
-            ALL THE CLOUDS, <br /> ALL AT ONCE.
-          </p>
-          <p className="text-lg md:text-xl mb-8 font-bold text-gray-100">
-            May 27th, 2025 @ Orpheum Theatre
-          </p>
-          <Button
-            size="lg"
-            className="bg-green-500 hover:bg-green-600 text-black"
-            onClick={openLink}
-          >
-            Register Now <ChevronRight className="ml-2 h-4 w-4" />
-          </Button>
-          <StatBanner />
-        </div>
-        <div className="grid grid-cols-4 gap-4 mt-32 md:grid-cols-4 md:flex md:flex-row md:overflow-hidden md:whitespace-nowrap">
-          {providers.map((provider) => (
-            <a
-              href={provider.link}
-              key={provider.name}
-              target="_blank"
-              className="relative"
-            >
-              <img
-                src={provider.icon}
-                alt={provider.name}
-                className="w-20 h-auto object-cover"
-              />
-            </a>
-          ))}
-        </div>
-      </header> */}
-      {/* Previous Event Photos */}
-      <section className="py-20 bg-gray-900" id="highlights">
-        <div className="container mx-auto">
-          <h2 className="text-3xl md:text-4xl font-bold mb-12 text-center bg-clip-text text-transparent bg-gradient-to-r from-green-400 to-blue-500">
-            Previous Event Highlights
-          </h2>
-
-          <div className="w-full mx-auto px-4 md:py-6 py-4">
-            <div className="relative flex justify-center">
-              <Carousel
-                className="md:w-full w-[90%]"
-                opts={{
-                  align: "start",
-                  loop: true
-                }}
-              >
-                <CarouselContent className="-ml-4">
-                  {[1, 2, 3, 4, 5].map((index) => (
-                    <CarouselItem
-                      key={index}
-                      className="lg:basis-1/4 md:basis-1/3 sm:basis-1/2"
-                    >
-                      <div
-                        key={index}
-                        className="relative h-64 rounded-lg group"
-                      >
-                        <Image
-                          src={`/past-events/AWSDay-${index}.jpg`}
-                          alt={`Event Photo ${index}`}
-                          layout="fill"
-                          objectFit="cover"
-                          className="transition-transform duration-300 group-hover:scale-110"
-                        />
-                        <div className="absolute inset-0 bg-gradient-to-t from-black/70 to-transparent opacity-0 group-hover:opacity-100 transition-opacity duration-300 flex items-end p-4">
-                          <p className="text-white text-lg font-semibold">
-                            Cloud Summit 2024
-                          </p>
-                        </div>
-                      </div>
-                    </CarouselItem>
-                  ))}
-                </CarouselContent>
-                <CarouselPrevious className="absolute lg:-left-5 -left-7 top-1/2 transform -translate-y-1/2 bg-primary/90 hover:bg-primary text-white border border-white p-4" />
-                <CarouselNext className="absolute lg:-right-5 -right-7 top-1/2 transform -translate-y-1/2 bg-primary/90 hover:bg-primary text-white rounded-full border border-white p-4" />
-              </Carousel>
-            </div>
-          </div>
-        </div>
-      </section>
-      {/* Sponsor Logos */}
-      <section className="py-20 bg-black">
-        <Sponsors />
-      </section>
-      {/* Call for Speakers and Volunteers */}
-      <section className="py-20 bg-gradient-to-b from-gray-900 to-black">
-        <div className="container mx-auto px-4 text-center">
-          <h2 className="text-3xl md:text-4xl font-bold mb-12 bg-clip-text text-transparent bg-gradient-to-r from-green-400 to-blue-500">
-            Join Our Community
-          </h2>
-          <div className="grid md:grid-cols-2 gap-8">
-            <Card className="bg-gray-800 border-gray-700">
-              <CardHeader>
-                <Megaphone className="w-12 h-12 mx-auto mb-4 text-green-500" />
-                <CardTitle className="text-2xl font-semibold">
-                  Call for Speakers
-                </CardTitle>
-                <CardDescription>
-                  Share your expertise with our community
-                </CardDescription>
-              </CardHeader>
-              <CardContent>
-                <p className="mb-4">
-                  Be part of shaping the future of cloud technology. Submit your
-                  talk proposal and inspire the next generation of innovators.
-                </p>
-              </CardContent>
-              <CardFooter className="justify-center">
-                <Button
-                  onClick={openSpeakersLink}
-                  variant="outline"
-                  className="border-green-500 text-green-500 hover:bg-green-500 hover:text-black"
-                >
-                  Apply Now
-                </Button>
-              </CardFooter>
-            </Card>
-            <Card className="bg-gray-800 border-gray-700">
-              <CardHeader>
-                <Users className="w-12 h-12 mx-auto mb-4 text-blue-500" />
-                <CardTitle className="text-2xl font-semibold">
-                  Volunteer Opportunities
-                </CardTitle>
-                <CardDescription>
-                  Help make Cloud Summit a success
-                </CardDescription>
-              </CardHeader>
-              <CardContent>
-                <p className="mb-4">
-                  Join our team of dedicated volunteers and play a crucial role
-                  in creating an unforgettable experience for all attendees.
-                </p>
-              </CardContent>
-              <CardFooter className="justify-center">
-                <Button
-                  onClick={openVolunteersLink}
-                  variant="outline"
-                  className="border-blue-500 text-blue-500 hover:bg-blue-500 hover:text-black"
-                >
-                  Join Us
-                </Button>
-              </CardFooter>
-            </Card>
-          </div>
-        </div>
-      </section>
-      {/* Venue Information */}
-      <section
-        className="py-20 bg-gradient-to-b from-black to-gray-900"
-        id="venue"
-      >
-        <div className="container mx-auto px-4 xl:max-w-[1100px]">
-          <h2 className="text-3xl md:text-4xl font-bold mb-12 text-center bg-clip-text text-transparent bg-gradient-to-r from-green-400 to-blue-500">
-            Venue Information
-          </h2>
-          <div className="flex flex-col md:flex-row items-center justify-center w-full gap-8">
-            <div className="md:w-max w-full">
-              <iframe
-                src="https://www.google.com/maps/embed?pb=!1m14!1m8!1m3!1d10411.22242662249!2d-123.1204164!3d49.2800806!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x5486717fcd9b875d%3A0x6f19bb45bb73289c!2sOrpheum!5e0!3m2!1sen!2sca!4v1729318806028!5m2!1sen!2sca"
-                className="h-[300px] md:w-[600px] w-full"
-                loading="lazy"
-              ></iframe>
-            </div>
-            <div className="md:w-1/3 space-y-6">
-              <Card className="bg-gray-800 border-gray-700">
-                <CardContent className="p-6">
-                  <div className="flex items-center mb-4">
-                    <Building className="w-6 h-6 mr-3 text-green-700" />
-                    <p className="text-lg">The Orpheum Theatre</p>
-                  </div>
-                  <div className="flex items-center mb-4">
-                    <MapPin className="w-6 h-6 mr-3 text-green-700" />
-                    <p className="text-lg">601 Smithe Street, Vancouver</p>
-                  </div>
-                  <div className="flex items-center">
-                    <CalendarDays className="w-6 h-6 mr-3 text-green-700" />
-                    <p className="text-lg">May 27th, 2025</p>
-                  </div>
-                </CardContent>
-              </Card>
-              <p className="text-gray-300">
-                Experience the Vancouver Cloud Summit like no technology
-                conference you have been to before. Sophistication and elegance
-                meets new age technology and engagement in an amazing venue in
-                the heart of downtown Vancouver. 2 large stages, many intimate
-                workshops and talks as well as a few surprises!
-              </p>
-            </div>
-          </div>
-        </div>
-      </section>
-      <section
-        className="py-20 bg-gradient-to-b from-gray-900 to-black"
-        id="faq"
-      >
-        <div className="container mx-auto px-4 xl:max-w-[1100px]">
-          <h2 className="text-3xl md:text-4xl font-bold mb-10 text-center bg-clip-text text-transparent bg-gradient-to-r from-green-400 to-blue-500">
-            FAQ
-          </h2>
-          <FAQaccordion />
-        </div>
-      </section>
+      <main>
+        <HeroSection />
+        <HighlightsSection />
+        <SponsorsSection />
+        <CommunitySection />
+        <VenueSection />
+        <FAQSection />
+      </main>
       <Footer />
     </div>
   );
